pyproject.toml
[tool.poetry]
name = "inkind-enterprise"
version = "0.1.0"
description = ""
authors = ["Natalia Ruemmele <natalia.ruemmele@inkind.at>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.4"
django-types = "^0.22.0"
django-stubs-ext = "^5.2.2"
requests = "^2.32.4"
beautifulsoup4 = "^4.13.4"
six = "^1.17.0"
python-dotenv = "^1.1.1"
pillow = "^11.3.0"
django-constance = "^4.3.2"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
django-storages = {extras = ["boto3"], version = "^1.14.6"}
boto3 = "^1.40.11"
django-axes = "^8.0.0"
argon2-cffi = "^25.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
django-debug-toolbar = "^4.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
